Fixes popups in ir_black for vim in terminal
Using ir_black in a terminal version of vim (i.e. not macvim) used to
show the popup background color as the same color as the regular
background. This caused things like fuzzyfinder textmate to not work
properly. Choosing the proper file was difficult because the colors were
the same.